db-importer/requirements.txt
# my_project/db-importer/requirements.txt
mysql-connector-python